refactor: subpath exports, delete infer/, extract activity types + root help
Breaking: split single entry into 3 subpaths (., ./testkit, ./runtime).
Test utilities (runCommand, createCaptureOutput, createTestPrompter,
createTestAdapter, PROMPT_CANCEL) move to dreamcli/testkit. Runtime
adapters move to dreamcli/runtime. Hard break, no deprecated re-exports.

- Delete empty infer/ stub + GOALS.md reference
- Extract 7 activity types from schema/command.ts to schema/activity.ts
  (Fallback, SpinnerHandle, ProgressHandle, SpinnerOptions, ProgressOptions,
  ActivityEvent, TableColumn). Out stays in command.ts
- Extract formatRootHelp + text helpers from cli/index.ts to cli/root-help.ts
- tsdown named entry object, exports:true auto-generates package.json
- attw ignoreRules: no-resolution (node10 can't resolve subpaths)
