--- src/core/infer/index.ts
@@ -1,7 +1,0 @@
-/**
- * TypeScript inference utility types for schema-to-handler type flow.
- *
- * @module dreamcli/core/infer
- */
-
-export {};
