+++ src/core/schema/command.ts
@@ -10,126 +10,19 @@
  */
 
 import type { RunOptions, RunResult } from '../testkit/index.js';
+import type {
+	ActivityEvent,
+	Fallback,
+	ProgressHandle,
+	ProgressOptions,
+	SpinnerHandle,
+	SpinnerOptions,
+	TableColumn,
+} from './activity.js';
 import type { ArgBuilder, ArgConfig, ArgSchema, InferArgs } from './arg.js';
 import type { FlagBuilder, FlagConfig, FlagSchema, InferFlags } from './flag.js';
 import type { ErasedMiddlewareHandler, Middleware } from './middleware.js';
 import type { PromptConfig } from './prompt.js';
-
-// ---------------------------------------------------------------------------
-// Activity types (spinner / progress)
-// ---------------------------------------------------------------------------
-
-/**
- * Non-TTY fallback strategy for spinners and progress bars.
- *
- * - `'silent'` — no output at all (default). Ideal for CI where decorative
- *   output is noise.
- * - `'static'` — emit plain text via `out.log()` / `out.error()` at
- *   lifecycle boundaries (start, succeed, fail). No animation.
- */
-type Fallback = 'silent' | 'static';
-
-/** Options for {@link Out.spinner}. */
-interface SpinnerOptions {
-	/** Fallback strategy when `!isTTY` or `jsonMode`. Defaults to `'silent'`. */
-	readonly fallback?: Fallback;
-}
-
-/**
- * Handle returned by {@link Out.spinner} for lifecycle control.
- *
- * Terminal methods (`succeed`, `fail`, `stop`) are idempotent — calling any
- * of them after the handle is already stopped is a no-op, not an error.
- */
-interface SpinnerHandle {
-	/** Update the spinner text (no-op if stopped). */
-	update(text: string): void;
-	/** Stop with a success symbol and optional final text. */
-	succeed(text?: string): void;
-	/** Stop with a failure symbol and optional final text. */
-	fail(text?: string): void;
-	/** Stop the spinner without a status symbol. */
-	stop(): void;
-	/**
-	 * Wrap a promise: auto-succeed on resolve, auto-fail on reject.
-	 *
-	 * @returns The resolved value of the wrapped promise.
-	 */
-	wrap<T>(
-		promise: Promise<T>,
-		options?: {
-			readonly succeed?: string;
-			readonly fail?: string;
-		},
-	): Promise<T>;
-}
-
-/** Options for {@link Out.progress}. */
-interface ProgressOptions {
-	/**
-	 * Total units of work. When provided, the bar shows a determinate
-	 * percentage. When omitted, the bar pulses in indeterminate mode.
-	 */
-	readonly total?: number;
-	/** Label displayed alongside the progress bar. */
-	readonly label?: string;
-	/** Fallback strategy when `!isTTY` or `jsonMode`. Defaults to `'silent'`. */
-	readonly fallback?: Fallback;
-}
-
-/**
- * Handle returned by {@link Out.progress} for lifecycle control.
- *
- * Terminal methods (`done`, `fail`) are idempotent — calling any
- * of them after the handle is already stopped is a no-op.
- */
-interface ProgressHandle {
-	/** Advance progress by `n` units (default 1). */
-	increment(n?: number): void;
-	/** Set progress to an absolute value. */
-	update(value: number): void;
-	/** Mark progress as complete with an optional final message. */
-	done(text?: string): void;
-	/** Mark progress as failed with an optional final message. */
-	fail(text?: string): void;
-}
-
-/**
- * Discriminated union of spinner and progress lifecycle events.
- *
- * Captured by testkit in {@link RunResult.activity} for assertion
- * without polluting stdout/stderr arrays.
- */
-type ActivityEvent =
-	| { readonly type: 'spinner:start'; readonly text: string }
-	| { readonly type: 'spinner:update'; readonly text: string }
-	| { readonly type: 'spinner:succeed'; readonly text: string }
-	| { readonly type: 'spinner:fail'; readonly text: string }
-	| { readonly type: 'spinner:stop' }
-	| { readonly type: 'progress:start'; readonly label: string; readonly total: number | undefined }
-	| { readonly type: 'progress:increment'; readonly delta: number }
-	| { readonly type: 'progress:update'; readonly value: number }
-	| { readonly type: 'progress:done'; readonly text: string | undefined }
-	| { readonly type: 'progress:fail'; readonly text: string | undefined };
-
-// ---------------------------------------------------------------------------
-// Table column descriptor
-// ---------------------------------------------------------------------------
-
-/**
- * Describes a single column in table output.
- *
- * @typeParam T - The row object type (inferred from the rows array).
- */
-interface TableColumn<T extends Record<string, unknown>> {
-	/** Property key on the row objects to display in this column. */
-	readonly key: keyof T & string;
-	/**
-	 * Header label for the column.
-	 * Defaults to the `key` value when omitted.
-	 */
-	readonly header?: string;
-}
 
 // ---------------------------------------------------------------------------
 // Context type utilities
@@ -877,7 +770,6 @@ export { command, group, CommandBuilder };
 export type {
 	ActionHandler,
 	ActionParams,
-	ActivityEvent,
 	AnyCommandBuilder,
 	CommandArgEntry,
 	CommandConfig,
@@ -885,14 +777,8 @@ export type {
 	CommandSchema,
 	ErasedCommand,
 	ErasedInteractiveResolver,
-	Fallback,
 	InteractiveParams,
 	InteractiveResolver,
 	InteractiveResult,
 	Out,
-	ProgressHandle,
-	ProgressOptions,
-	SpinnerHandle,
-	SpinnerOptions,
-	TableColumn,
 };
