+++ src/index.test.ts
@@ -7,11 +7,24 @@ describe('dreamcli', () => {
 	});
 });
 
+describe('dreamcli/testkit', () => {
+	it('module loads without error', async () => {
+		const mod = await import('./testkit.js');
+		expect(mod).toBeDefined();
+	});
+});
+
+describe('dreamcli/runtime', () => {
+	it('module loads without error', async () => {
+		const mod = await import('./runtime.js');
+		expect(mod).toBeDefined();
+	});
+});
+
 describe('project structure', () => {
 	const coreModules = [
 		'./core/errors/index.js',
 		'./core/schema/index.js',
-		'./core/infer/index.js',
 		'./core/parse/index.js',
 		'./core/resolve/index.js',
 		'./core/help/index.js',
