+++ tsconfig.json
@@ -18,7 +18,6 @@
 		"declarationMap": true,
 		"sourceMap": true,
 		"outDir": "dist",
-		"rootDir": "src",
 
 		// Interop
 		"esModuleInterop": true,
